// client/src/components/simulators/PinTerminal/src/enums/SupportedSchemes.ts
export enum SupportedSchemesType {
  ALIPAY = 'Alipay',
  AMEX = 'AMEX',
  APPLEPAY = 'ApplePay',
  BANCONTACT = 'Bancontact',

  CARTESBANCAIRES = 'Cartes Bancaires',
  DINERS = 'Diners Club',
  DISCOVER = 'Discover',
  GIROCARD = 'Girocard',

  GIROPAY = 'Giropay',
  GOOGLEPAY = 'Google Pay',
  IDEAL = 'iDEAL',
  INTERAC = 'Interac',

  JCB_BANK = 'JCB',
  MAESTRO = 'Maestro',
  MASTERCARD = 'Mastercard',
  MASTERCARDDEBIT = 'Mastercard Debit',

  MASTERPASS = 'Masterpass',
  PAYPAL = 'PayPal',
  SWISSPOST = 'Swiss Post',
  SWISSREKA = 'Swiss Reka',

  TWINT = 'Twint',
  UNIONPAY = 'UnionPay',
  VISA = 'Visa',
  VISAELECTRON = 'Visa Electron',

  VISADEBIT = 'Visa Debit',
  VPAY = 'V PAY',
  WECHATPAY = 'WeChat Pay',
}